mysql: fix auth method switch to caching_sha2_password
A switch *to* caching_sha2_password can happen if a MySQL 8
server's default authentication method is changed but users
still use caching_sha2_password authentication.

This changes the password-scrambling operation behind
caching_sha2_password to trim its "scramble" (nonce) to
the appropriate length, like the other functions. (A "\0"
is included after the nonce on auth method switch
messages, but not in the original handshake message.)