mysql: update connection charset to utf8mb4
This changes connections to use "utf8mb4" instead of "utf8" (MySQL's
nonstandard subset of UTF-8) when possible. See also the following:
https://mysqlserverteam.com/mysql-8-0-when-to-use-utf8mb3-over-utf8mb4/

Complication: the wire protocol needs a *collation* identifier, not
just a charset, different versions of MySQL support different charsets
and collations, and the default collation for a charset might change
(eg from 5.7 to 8.0, maybe also via server configuration).

- 4.1 through 5.5.2
  - utf8 exists but utf8mb4 does not
- 5.5.3 (2010) through 5.7.*:
  - utf8mb4 exists, default collation is utf8mb4_general_ci
- since 8.0:
  - utf8mb4 exists, default collation is utf8mb4_0900_ai_ci

Previously, this library did "SET NAMES utf8" at the beginning of a
connection. We could change that to "SET NAMES utf8mb4", but that
would fail on old versions, and it would also be good to eliminate the
extra round trip if possible. So we do the following:

1. If the server's default collation is utf8mb4_general_ci or
utf8mb4_0900_ai_ci, then we use that.
2. If the server's version is 4.1 through 5.5.2, use utf8_general_ci.
3. Otherwise, use utf8mb4_general_ci.

If the client is running queries that are sensitive to the connection
collation, then it should use "SET NAMES" or "SET
collation_connection" (but it must only use collations whose charsets
are utf8mb4 or utf8).

Other references:
http://mysqlserverteam.com/sushi-beer-an-introduction-of-utf8-support-in-mysql-8-0/
http://mysqlserverteam.com/new-collations-in-mysql-8-0-0/