Add tests and docs for `prepare` accepting a virtual statement
The actual implementation of `prepare` is in base, so the version
constraint on base is also updated. Even though no code has changed in
db-lib, this allows a lower bound on db-lib to guarantee that the
functionality will be available (and that version can be listed in the
documentation).

See also racket/racket#5093, which includes the implementation part of
this change.